Enum av_sys::AVCodecID

#[repr(u32)]

pub enum AVCodecID {
    AV_CODEC_ID_NONE,
    AV_CODEC_ID_MPEG1VIDEO,
    AV_CODEC_ID_MPEG2VIDEO,
    AV_CODEC_ID_MPEG2VIDEO_XVMC,
    AV_CODEC_ID_H261,
    AV_CODEC_ID_H263,
    AV_CODEC_ID_RV10,
    AV_CODEC_ID_RV20,
    AV_CODEC_ID_MJPEG,
    AV_CODEC_ID_MJPEGB,
    AV_CODEC_ID_LJPEG,
    AV_CODEC_ID_SP5X,
    AV_CODEC_ID_JPEGLS,
    AV_CODEC_ID_MPEG4,
    AV_CODEC_ID_RAWVIDEO,
    AV_CODEC_ID_MSMPEG4V1,
    AV_CODEC_ID_MSMPEG4V2,
    AV_CODEC_ID_MSMPEG4V3,
    AV_CODEC_ID_WMV1,
    AV_CODEC_ID_WMV2,
    AV_CODEC_ID_H263P,
    AV_CODEC_ID_H263I,
    AV_CODEC_ID_FLV1,
    AV_CODEC_ID_SVQ1,
    AV_CODEC_ID_SVQ3,
    AV_CODEC_ID_DVVIDEO,
    AV_CODEC_ID_HUFFYUV,
    AV_CODEC_ID_CYUV,
    AV_CODEC_ID_H264,
    AV_CODEC_ID_INDEO3,
    AV_CODEC_ID_VP3,
    AV_CODEC_ID_THEORA,
    AV_CODEC_ID_ASV1,
    AV_CODEC_ID_ASV2,
    AV_CODEC_ID_FFV1,
    AV_CODEC_ID_4XM,
    AV_CODEC_ID_VCR1,
    AV_CODEC_ID_CLJR,
    AV_CODEC_ID_MDEC,
    AV_CODEC_ID_ROQ,
    AV_CODEC_ID_INTERPLAY_VIDEO,
    AV_CODEC_ID_XAN_WC3,
    AV_CODEC_ID_XAN_WC4,
    AV_CODEC_ID_RPZA,
    AV_CODEC_ID_CINEPAK,
    AV_CODEC_ID_WS_VQA,
    AV_CODEC_ID_MSRLE,
    AV_CODEC_ID_MSVIDEO1,
    AV_CODEC_ID_IDCIN,
    AV_CODEC_ID_8BPS,
    AV_CODEC_ID_SMC,
    AV_CODEC_ID_FLIC,
    AV_CODEC_ID_TRUEMOTION1,
    AV_CODEC_ID_VMDVIDEO,
    AV_CODEC_ID_MSZH,
    AV_CODEC_ID_ZLIB,
    AV_CODEC_ID_QTRLE,
    AV_CODEC_ID_TSCC,
    AV_CODEC_ID_ULTI,
    AV_CODEC_ID_QDRAW,
    AV_CODEC_ID_VIXL,
    AV_CODEC_ID_QPEG,
    AV_CODEC_ID_PNG,
    AV_CODEC_ID_PPM,
    AV_CODEC_ID_PBM,
    AV_CODEC_ID_PGM,
    AV_CODEC_ID_PGMYUV,
    AV_CODEC_ID_PAM,
    AV_CODEC_ID_FFVHUFF,
    AV_CODEC_ID_RV30,
    AV_CODEC_ID_RV40,
    AV_CODEC_ID_VC1,
    AV_CODEC_ID_WMV3,
    AV_CODEC_ID_LOCO,
    AV_CODEC_ID_WNV1,
    AV_CODEC_ID_AASC,
    AV_CODEC_ID_INDEO2,
    AV_CODEC_ID_FRAPS,
    AV_CODEC_ID_TRUEMOTION2,
    AV_CODEC_ID_BMP,
    AV_CODEC_ID_CSCD,
    AV_CODEC_ID_MMVIDEO,
    AV_CODEC_ID_ZMBV,
    AV_CODEC_ID_AVS,
    AV_CODEC_ID_SMACKVIDEO,
    AV_CODEC_ID_NUV,
    AV_CODEC_ID_KMVC,
    AV_CODEC_ID_FLASHSV,
    AV_CODEC_ID_CAVS,
    AV_CODEC_ID_JPEG2000,
    AV_CODEC_ID_VMNC,
    AV_CODEC_ID_VP5,
    AV_CODEC_ID_VP6,
    AV_CODEC_ID_VP6F,
    AV_CODEC_ID_TARGA,
    AV_CODEC_ID_DSICINVIDEO,
    AV_CODEC_ID_TIERTEXSEQVIDEO,
    AV_CODEC_ID_TIFF,
    AV_CODEC_ID_GIF,
    AV_CODEC_ID_DXA,
    AV_CODEC_ID_DNXHD,
    AV_CODEC_ID_THP,
    AV_CODEC_ID_SGI,
    AV_CODEC_ID_C93,
    AV_CODEC_ID_BETHSOFTVID,
    AV_CODEC_ID_PTX,
    AV_CODEC_ID_TXD,
    AV_CODEC_ID_VP6A,
    AV_CODEC_ID_AMV,
    AV_CODEC_ID_VB,
    AV_CODEC_ID_PCX,
    AV_CODEC_ID_SUNRAST,
    AV_CODEC_ID_INDEO4,
    AV_CODEC_ID_INDEO5,
    AV_CODEC_ID_MIMIC,
    AV_CODEC_ID_RL2,
    AV_CODEC_ID_ESCAPE124,
    AV_CODEC_ID_DIRAC,
    AV_CODEC_ID_BFI,
    AV_CODEC_ID_CMV,
    AV_CODEC_ID_MOTIONPIXELS,
    AV_CODEC_ID_TGV,
    AV_CODEC_ID_TGQ,
    AV_CODEC_ID_TQI,
    AV_CODEC_ID_AURA,
    AV_CODEC_ID_AURA2,
    AV_CODEC_ID_V210X,
    AV_CODEC_ID_TMV,
    AV_CODEC_ID_V210,
    AV_CODEC_ID_DPX,
    AV_CODEC_ID_MAD,
    AV_CODEC_ID_FRWU,
    AV_CODEC_ID_FLASHSV2,
    AV_CODEC_ID_CDGRAPHICS,
    AV_CODEC_ID_R210,
    AV_CODEC_ID_ANM,
    AV_CODEC_ID_BINKVIDEO,
    AV_CODEC_ID_IFF_ILBM,
    AV_CODEC_ID_KGV1,
    AV_CODEC_ID_YOP,
    AV_CODEC_ID_VP8,
    AV_CODEC_ID_PICTOR,
    AV_CODEC_ID_ANSI,
    AV_CODEC_ID_A64_MULTI,
    AV_CODEC_ID_A64_MULTI5,
    AV_CODEC_ID_R10K,
    AV_CODEC_ID_MXPEG,
    AV_CODEC_ID_LAGARITH,
    AV_CODEC_ID_PRORES,
    AV_CODEC_ID_JV,
    AV_CODEC_ID_DFA,
    AV_CODEC_ID_WMV3IMAGE,
    AV_CODEC_ID_VC1IMAGE,
    AV_CODEC_ID_UTVIDEO,
    AV_CODEC_ID_BMV_VIDEO,
    AV_CODEC_ID_VBLE,
    AV_CODEC_ID_DXTORY,
    AV_CODEC_ID_V410,
    AV_CODEC_ID_XWD,
    AV_CODEC_ID_CDXL,
    AV_CODEC_ID_XBM,
    AV_CODEC_ID_ZEROCODEC,
    AV_CODEC_ID_MSS1,
    AV_CODEC_ID_MSA1,
    AV_CODEC_ID_TSCC2,
    AV_CODEC_ID_MTS2,
    AV_CODEC_ID_CLLC,
    AV_CODEC_ID_MSS2,
    AV_CODEC_ID_VP9,
    AV_CODEC_ID_AIC,
    AV_CODEC_ID_ESCAPE130,
    AV_CODEC_ID_G2M,
    AV_CODEC_ID_WEBP,
    AV_CODEC_ID_HNM4_VIDEO,
    AV_CODEC_ID_HEVC,
    AV_CODEC_ID_FIC,
    AV_CODEC_ID_ALIAS_PIX,
    AV_CODEC_ID_BRENDER_PIX,
    AV_CODEC_ID_PAF_VIDEO,
    AV_CODEC_ID_EXR,
    AV_CODEC_ID_VP7,
    AV_CODEC_ID_SANM,
    AV_CODEC_ID_SGIRLE,
    AV_CODEC_ID_MVC1,
    AV_CODEC_ID_MVC2,
    AV_CODEC_ID_HQX,
    AV_CODEC_ID_TDSC,
    AV_CODEC_ID_HQ_HQA,
    AV_CODEC_ID_HAP,
    AV_CODEC_ID_DDS,
    AV_CODEC_ID_DXV,
    AV_CODEC_ID_SCREENPRESSO,
    AV_CODEC_ID_RSCC,
    AV_CODEC_ID_Y41P,
    AV_CODEC_ID_AVRP,
    AV_CODEC_ID_012V,
    AV_CODEC_ID_AVUI,
    AV_CODEC_ID_AYUV,
    AV_CODEC_ID_TARGA_Y216,
    AV_CODEC_ID_V308,
    AV_CODEC_ID_V408,
    AV_CODEC_ID_YUV4,
    AV_CODEC_ID_AVRN,
    AV_CODEC_ID_CPIA,
    AV_CODEC_ID_XFACE,
    AV_CODEC_ID_SNOW,
    AV_CODEC_ID_SMVJPEG,
    AV_CODEC_ID_APNG,
    AV_CODEC_ID_DAALA,
    AV_CODEC_ID_CFHD,
    AV_CODEC_ID_TRUEMOTION2RT,
    AV_CODEC_ID_M101,
    AV_CODEC_ID_MAGICYUV,
    AV_CODEC_ID_SHEERVIDEO,
    AV_CODEC_ID_YLC,
    AV_CODEC_ID_PSD,
    AV_CODEC_ID_PIXLET,
    AV_CODEC_ID_SPEEDHQ,
    AV_CODEC_ID_FMVC,
    AV_CODEC_ID_SCPR,
    AV_CODEC_ID_CLEARVIDEO,
    AV_CODEC_ID_XPM,
    AV_CODEC_ID_AV1,
    AV_CODEC_ID_FIRST_AUDIO,
    AV_CODEC_ID_PCM_S16BE,
    AV_CODEC_ID_PCM_U16LE,
    AV_CODEC_ID_PCM_U16BE,
    AV_CODEC_ID_PCM_S8,
    AV_CODEC_ID_PCM_U8,
    AV_CODEC_ID_PCM_MULAW,
    AV_CODEC_ID_PCM_ALAW,
    AV_CODEC_ID_PCM_S32LE,
    AV_CODEC_ID_PCM_S32BE,
    AV_CODEC_ID_PCM_U32LE,
    AV_CODEC_ID_PCM_U32BE,
    AV_CODEC_ID_PCM_S24LE,
    AV_CODEC_ID_PCM_S24BE,
    AV_CODEC_ID_PCM_U24LE,
    AV_CODEC_ID_PCM_U24BE,
    AV_CODEC_ID_PCM_S24DAUD,
    AV_CODEC_ID_PCM_ZORK,
    AV_CODEC_ID_PCM_S16LE_PLANAR,
    AV_CODEC_ID_PCM_DVD,
    AV_CODEC_ID_PCM_F32BE,
    AV_CODEC_ID_PCM_F32LE,
    AV_CODEC_ID_PCM_F64BE,
    AV_CODEC_ID_PCM_F64LE,
    AV_CODEC_ID_PCM_BLURAY,
    AV_CODEC_ID_PCM_LXF,
    AV_CODEC_ID_S302M,
    AV_CODEC_ID_PCM_S8_PLANAR,
    AV_CODEC_ID_PCM_S24LE_PLANAR,
    AV_CODEC_ID_PCM_S32LE_PLANAR,
    AV_CODEC_ID_PCM_S16BE_PLANAR,
    AV_CODEC_ID_PCM_S64LE,
    AV_CODEC_ID_PCM_S64BE,
    AV_CODEC_ID_PCM_F16LE,
    AV_CODEC_ID_PCM_F24LE,
    AV_CODEC_ID_ADPCM_IMA_QT,
    AV_CODEC_ID_ADPCM_IMA_WAV,
    AV_CODEC_ID_ADPCM_IMA_DK3,
    AV_CODEC_ID_ADPCM_IMA_DK4,
    AV_CODEC_ID_ADPCM_IMA_WS,
    AV_CODEC_ID_ADPCM_IMA_SMJPEG,
    AV_CODEC_ID_ADPCM_MS,
    AV_CODEC_ID_ADPCM_4XM,
    AV_CODEC_ID_ADPCM_XA,
    AV_CODEC_ID_ADPCM_ADX,
    AV_CODEC_ID_ADPCM_EA,
    AV_CODEC_ID_ADPCM_G726,
    AV_CODEC_ID_ADPCM_CT,
    AV_CODEC_ID_ADPCM_SWF,
    AV_CODEC_ID_ADPCM_YAMAHA,
    AV_CODEC_ID_ADPCM_SBPRO_4,
    AV_CODEC_ID_ADPCM_SBPRO_3,
    AV_CODEC_ID_ADPCM_SBPRO_2,
    AV_CODEC_ID_ADPCM_THP,
    AV_CODEC_ID_ADPCM_IMA_AMV,
    AV_CODEC_ID_ADPCM_EA_R1,
    AV_CODEC_ID_ADPCM_EA_R3,
    AV_CODEC_ID_ADPCM_EA_R2,
    AV_CODEC_ID_ADPCM_IMA_EA_SEAD,
    AV_CODEC_ID_ADPCM_IMA_EA_EACS,
    AV_CODEC_ID_ADPCM_EA_XAS,
    AV_CODEC_ID_ADPCM_EA_MAXIS_XA,
    AV_CODEC_ID_ADPCM_IMA_ISS,
    AV_CODEC_ID_ADPCM_G722,
    AV_CODEC_ID_ADPCM_IMA_APC,
    AV_CODEC_ID_ADPCM_VIMA,
    AV_CODEC_ID_ADPCM_AFC,
    AV_CODEC_ID_ADPCM_IMA_OKI,
    AV_CODEC_ID_ADPCM_DTK,
    AV_CODEC_ID_ADPCM_IMA_RAD,
    AV_CODEC_ID_ADPCM_G726LE,
    AV_CODEC_ID_ADPCM_THP_LE,
    AV_CODEC_ID_ADPCM_PSX,
    AV_CODEC_ID_ADPCM_AICA,
    AV_CODEC_ID_ADPCM_IMA_DAT4,
    AV_CODEC_ID_ADPCM_MTAF,
    AV_CODEC_ID_AMR_NB,
    AV_CODEC_ID_AMR_WB,
    AV_CODEC_ID_RA_144,
    AV_CODEC_ID_RA_288,
    AV_CODEC_ID_ROQ_DPCM,
    AV_CODEC_ID_INTERPLAY_DPCM,
    AV_CODEC_ID_XAN_DPCM,
    AV_CODEC_ID_SOL_DPCM,
    AV_CODEC_ID_SDX2_DPCM,
    AV_CODEC_ID_MP2,
    AV_CODEC_ID_MP3,
    AV_CODEC_ID_AAC,
    AV_CODEC_ID_AC3,
    AV_CODEC_ID_DTS,
    AV_CODEC_ID_VORBIS,
    AV_CODEC_ID_DVAUDIO,
    AV_CODEC_ID_WMAV1,
    AV_CODEC_ID_WMAV2,
    AV_CODEC_ID_MACE3,
    AV_CODEC_ID_MACE6,
    AV_CODEC_ID_VMDAUDIO,
    AV_CODEC_ID_FLAC,
    AV_CODEC_ID_MP3ADU,
    AV_CODEC_ID_MP3ON4,
    AV_CODEC_ID_SHORTEN,
    AV_CODEC_ID_ALAC,
    AV_CODEC_ID_WESTWOOD_SND1,
    AV_CODEC_ID_GSM,
    AV_CODEC_ID_QDM2,
    AV_CODEC_ID_COOK,
    AV_CODEC_ID_TRUESPEECH,
    AV_CODEC_ID_TTA,
    AV_CODEC_ID_SMACKAUDIO,
    AV_CODEC_ID_QCELP,
    AV_CODEC_ID_WAVPACK,
    AV_CODEC_ID_DSICINAUDIO,
    AV_CODEC_ID_IMC,
    AV_CODEC_ID_MUSEPACK7,
    AV_CODEC_ID_MLP,
    AV_CODEC_ID_GSM_MS,
    AV_CODEC_ID_ATRAC3,
    AV_CODEC_ID_VOXWARE,
    AV_CODEC_ID_APE,
    AV_CODEC_ID_NELLYMOSER,
    AV_CODEC_ID_MUSEPACK8,
    AV_CODEC_ID_SPEEX,
    AV_CODEC_ID_WMAVOICE,
    AV_CODEC_ID_WMAPRO,
    AV_CODEC_ID_WMALOSSLESS,
    AV_CODEC_ID_ATRAC3P,
    AV_CODEC_ID_EAC3,
    AV_CODEC_ID_SIPR,
    AV_CODEC_ID_MP1,
    AV_CODEC_ID_TWINVQ,
    AV_CODEC_ID_TRUEHD,
    AV_CODEC_ID_MP4ALS,
    AV_CODEC_ID_ATRAC1,
    AV_CODEC_ID_BINKAUDIO_RDFT,
    AV_CODEC_ID_BINKAUDIO_DCT,
    AV_CODEC_ID_AAC_LATM,
    AV_CODEC_ID_QDMC,
    AV_CODEC_ID_CELT,
    AV_CODEC_ID_G723_1,
    AV_CODEC_ID_G729,
    AV_CODEC_ID_8SVX_EXP,
    AV_CODEC_ID_8SVX_FIB,
    AV_CODEC_ID_BMV_AUDIO,
    AV_CODEC_ID_RALF,
    AV_CODEC_ID_IAC,
    AV_CODEC_ID_ILBC,
    AV_CODEC_ID_OPUS,
    AV_CODEC_ID_COMFORT_NOISE,
    AV_CODEC_ID_TAK,
    AV_CODEC_ID_METASOUND,
    AV_CODEC_ID_PAF_AUDIO,
    AV_CODEC_ID_ON2AVC,
    AV_CODEC_ID_DSS_SP,
    AV_CODEC_ID_FFWAVESYNTH,
    AV_CODEC_ID_SONIC,
    AV_CODEC_ID_SONIC_LS,
    AV_CODEC_ID_EVRC,
    AV_CODEC_ID_SMV,
    AV_CODEC_ID_DSD_LSBF,
    AV_CODEC_ID_DSD_MSBF,
    AV_CODEC_ID_DSD_LSBF_PLANAR,
    AV_CODEC_ID_DSD_MSBF_PLANAR,
    AV_CODEC_ID_4GV,
    AV_CODEC_ID_INTERPLAY_ACM,
    AV_CODEC_ID_XMA1,
    AV_CODEC_ID_XMA2,
    AV_CODEC_ID_DST,
    AV_CODEC_ID_ATRAC3AL,
    AV_CODEC_ID_ATRAC3PAL,
    AV_CODEC_ID_FIRST_SUBTITLE,
    AV_CODEC_ID_DVB_SUBTITLE,
    AV_CODEC_ID_TEXT,
    AV_CODEC_ID_XSUB,
    AV_CODEC_ID_SSA,
    AV_CODEC_ID_MOV_TEXT,
    AV_CODEC_ID_HDMV_PGS_SUBTITLE,
    AV_CODEC_ID_DVB_TELETEXT,
    AV_CODEC_ID_SRT,
    AV_CODEC_ID_MICRODVD,
    AV_CODEC_ID_EIA_608,
    AV_CODEC_ID_JACOSUB,
    AV_CODEC_ID_SAMI,
    AV_CODEC_ID_REALTEXT,
    AV_CODEC_ID_STL,
    AV_CODEC_ID_SUBVIEWER1,
    AV_CODEC_ID_SUBVIEWER,
    AV_CODEC_ID_SUBRIP,
    AV_CODEC_ID_WEBVTT,
    AV_CODEC_ID_MPL2,
    AV_CODEC_ID_VPLAYER,
    AV_CODEC_ID_PJS,
    AV_CODEC_ID_ASS,
    AV_CODEC_ID_HDMV_TEXT_SUBTITLE,
    AV_CODEC_ID_FIRST_UNKNOWN,
    AV_CODEC_ID_SCTE_35,
    AV_CODEC_ID_BINTEXT,
    AV_CODEC_ID_XBIN,
    AV_CODEC_ID_IDF,
    AV_CODEC_ID_OTF,
    AV_CODEC_ID_SMPTE_KLV,
    AV_CODEC_ID_DVD_NAV,
    AV_CODEC_ID_TIMED_ID3,
    AV_CODEC_ID_BIN_DATA,
    AV_CODEC_ID_PROBE,
    AV_CODEC_ID_MPEG2TS,
    AV_CODEC_ID_MPEG4SYSTEMS,
    AV_CODEC_ID_FFMETADATA,
    AV_CODEC_ID_WRAPPED_AVFRAME,
}

Identify the syntax and semantics of the bitstream. The principle is roughly: Two decoders with the same ID can decode the same streams. Two encoders with the same ID can encode compatible streams. There may be slight deviations from the principle due to implementation details.

If you add a codec ID to this list, add it so that 1. no value of an existing codec ID changes (that would break ABI), 2. it is as close as possible to similar codecs

After adding new codec IDs, do not forget to add an entry to the codec descriptor list and bump libavcodec minor version.

Variants
AV_CODEC_ID_NONEAV_CODEC_ID_MPEG1VIDEOAV_CODEC_ID_MPEG2VIDEOAV_CODEC_ID_MPEG2VIDEO_XVMCAV_CODEC_ID_H261AV_CODEC_ID_H263AV_CODEC_ID_RV10AV_CODEC_ID_RV20AV_CODEC_ID_MJPEGAV_CODEC_ID_MJPEGBAV_CODEC_ID_LJPEGAV_CODEC_ID_SP5XAV_CODEC_ID_JPEGLSAV_CODEC_ID_MPEG4AV_CODEC_ID_RAWVIDEOAV_CODEC_ID_MSMPEG4V1AV_CODEC_ID_MSMPEG4V2AV_CODEC_ID_MSMPEG4V3AV_CODEC_ID_WMV1AV_CODEC_ID_WMV2AV_CODEC_ID_H263PAV_CODEC_ID_H263IAV_CODEC_ID_FLV1AV_CODEC_ID_SVQ1AV_CODEC_ID_SVQ3AV_CODEC_ID_DVVIDEOAV_CODEC_ID_HUFFYUVAV_CODEC_ID_CYUVAV_CODEC_ID_H264AV_CODEC_ID_INDEO3AV_CODEC_ID_VP3AV_CODEC_ID_THEORAAV_CODEC_ID_ASV1AV_CODEC_ID_ASV2AV_CODEC_ID_FFV1AV_CODEC_ID_4XMAV_CODEC_ID_VCR1AV_CODEC_ID_CLJRAV_CODEC_ID_MDECAV_CODEC_ID_ROQAV_CODEC_ID_INTERPLAY_VIDEOAV_CODEC_ID_XAN_WC3AV_CODEC_ID_XAN_WC4AV_CODEC_ID_RPZAAV_CODEC_ID_CINEPAKAV_CODEC_ID_WS_VQAAV_CODEC_ID_MSRLEAV_CODEC_ID_MSVIDEO1AV_CODEC_ID_IDCINAV_CODEC_ID_8BPSAV_CODEC_ID_SMCAV_CODEC_ID_FLICAV_CODEC_ID_TRUEMOTION1AV_CODEC_ID_VMDVIDEOAV_CODEC_ID_MSZHAV_CODEC_ID_ZLIBAV_CODEC_ID_QTRLEAV_CODEC_ID_TSCCAV_CODEC_ID_ULTIAV_CODEC_ID_QDRAWAV_CODEC_ID_VIXLAV_CODEC_ID_QPEGAV_CODEC_ID_PNGAV_CODEC_ID_PPMAV_CODEC_ID_PBMAV_CODEC_ID_PGMAV_CODEC_ID_PGMYUVAV_CODEC_ID_PAMAV_CODEC_ID_FFVHUFFAV_CODEC_ID_RV30AV_CODEC_ID_RV40AV_CODEC_ID_VC1AV_CODEC_ID_WMV3AV_CODEC_ID_LOCOAV_CODEC_ID_WNV1AV_CODEC_ID_AASCAV_CODEC_ID_INDEO2AV_CODEC_ID_FRAPSAV_CODEC_ID_TRUEMOTION2AV_CODEC_ID_BMPAV_CODEC_ID_CSCDAV_CODEC_ID_MMVIDEOAV_CODEC_ID_ZMBVAV_CODEC_ID_AVSAV_CODEC_ID_SMACKVIDEOAV_CODEC_ID_NUVAV_CODEC_ID_KMVCAV_CODEC_ID_FLASHSVAV_CODEC_ID_CAVSAV_CODEC_ID_JPEG2000AV_CODEC_ID_VMNCAV_CODEC_ID_VP5AV_CODEC_ID_VP6AV_CODEC_ID_VP6FAV_CODEC_ID_TARGAAV_CODEC_ID_DSICINVIDEOAV_CODEC_ID_TIERTEXSEQVIDEOAV_CODEC_ID_TIFFAV_CODEC_ID_GIFAV_CODEC_ID_DXAAV_CODEC_ID_DNXHDAV_CODEC_ID_THPAV_CODEC_ID_SGIAV_CODEC_ID_C93AV_CODEC_ID_BETHSOFTVIDAV_CODEC_ID_PTXAV_CODEC_ID_TXDAV_CODEC_ID_VP6AAV_CODEC_ID_AMVAV_CODEC_ID_VBAV_CODEC_ID_PCXAV_CODEC_ID_SUNRASTAV_CODEC_ID_INDEO4AV_CODEC_ID_INDEO5AV_CODEC_ID_MIMICAV_CODEC_ID_RL2AV_CODEC_ID_ESCAPE124AV_CODEC_ID_DIRACAV_CODEC_ID_BFIAV_CODEC_ID_CMVAV_CODEC_ID_MOTIONPIXELSAV_CODEC_ID_TGVAV_CODEC_ID_TGQAV_CODEC_ID_TQIAV_CODEC_ID_AURAAV_CODEC_ID_AURA2AV_CODEC_ID_V210XAV_CODEC_ID_TMVAV_CODEC_ID_V210AV_CODEC_ID_DPXAV_CODEC_ID_MADAV_CODEC_ID_FRWUAV_CODEC_ID_FLASHSV2AV_CODEC_ID_CDGRAPHICSAV_CODEC_ID_R210AV_CODEC_ID_ANMAV_CODEC_ID_BINKVIDEOAV_CODEC_ID_IFF_ILBMAV_CODEC_ID_KGV1AV_CODEC_ID_YOPAV_CODEC_ID_VP8AV_CODEC_ID_PICTORAV_CODEC_ID_ANSIAV_CODEC_ID_A64_MULTIAV_CODEC_ID_A64_MULTI5AV_CODEC_ID_R10KAV_CODEC_ID_MXPEGAV_CODEC_ID_LAGARITHAV_CODEC_ID_PRORESAV_CODEC_ID_JVAV_CODEC_ID_DFAAV_CODEC_ID_WMV3IMAGEAV_CODEC_ID_VC1IMAGEAV_CODEC_ID_UTVIDEOAV_CODEC_ID_BMV_VIDEOAV_CODEC_ID_VBLEAV_CODEC_ID_DXTORYAV_CODEC_ID_V410AV_CODEC_ID_XWDAV_CODEC_ID_CDXLAV_CODEC_ID_XBMAV_CODEC_ID_ZEROCODECAV_CODEC_ID_MSS1AV_CODEC_ID_MSA1AV_CODEC_ID_TSCC2AV_CODEC_ID_MTS2AV_CODEC_ID_CLLCAV_CODEC_ID_MSS2AV_CODEC_ID_VP9AV_CODEC_ID_AICAV_CODEC_ID_ESCAPE130AV_CODEC_ID_G2MAV_CODEC_ID_WEBPAV_CODEC_ID_HNM4_VIDEOAV_CODEC_ID_HEVCAV_CODEC_ID_FICAV_CODEC_ID_ALIAS_PIXAV_CODEC_ID_BRENDER_PIXAV_CODEC_ID_PAF_VIDEOAV_CODEC_ID_EXRAV_CODEC_ID_VP7AV_CODEC_ID_SANMAV_CODEC_ID_SGIRLEAV_CODEC_ID_MVC1AV_CODEC_ID_MVC2AV_CODEC_ID_HQXAV_CODEC_ID_TDSCAV_CODEC_ID_HQ_HQAAV_CODEC_ID_HAPAV_CODEC_ID_DDSAV_CODEC_ID_DXVAV_CODEC_ID_SCREENPRESSOAV_CODEC_ID_RSCCAV_CODEC_ID_Y41PAV_CODEC_ID_AVRPAV_CODEC_ID_012VAV_CODEC_ID_AVUIAV_CODEC_ID_AYUVAV_CODEC_ID_TARGA_Y216AV_CODEC_ID_V308AV_CODEC_ID_V408AV_CODEC_ID_YUV4AV_CODEC_ID_AVRNAV_CODEC_ID_CPIAAV_CODEC_ID_XFACEAV_CODEC_ID_SNOWAV_CODEC_ID_SMVJPEGAV_CODEC_ID_APNGAV_CODEC_ID_DAALAAV_CODEC_ID_CFHDAV_CODEC_ID_TRUEMOTION2RTAV_CODEC_ID_M101AV_CODEC_ID_MAGICYUVAV_CODEC_ID_SHEERVIDEOAV_CODEC_ID_YLCAV_CODEC_ID_PSDAV_CODEC_ID_PIXLETAV_CODEC_ID_SPEEDHQAV_CODEC_ID_FMVCAV_CODEC_ID_SCPRAV_CODEC_ID_CLEARVIDEOAV_CODEC_ID_XPMAV_CODEC_ID_AV1AV_CODEC_ID_FIRST_AUDIOAV_CODEC_ID_PCM_S16BEAV_CODEC_ID_PCM_U16LEAV_CODEC_ID_PCM_U16BEAV_CODEC_ID_PCM_S8AV_CODEC_ID_PCM_U8AV_CODEC_ID_PCM_MULAWAV_CODEC_ID_PCM_ALAWAV_CODEC_ID_PCM_S32LEAV_CODEC_ID_PCM_S32BEAV_CODEC_ID_PCM_U32LEAV_CODEC_ID_PCM_U32BEAV_CODEC_ID_PCM_S24LEAV_CODEC_ID_PCM_S24BEAV_CODEC_ID_PCM_U24LEAV_CODEC_ID_PCM_U24BEAV_CODEC_ID_PCM_S24DAUDAV_CODEC_ID_PCM_ZORKAV_CODEC_ID_PCM_S16LE_PLANARAV_CODEC_ID_PCM_DVDAV_CODEC_ID_PCM_F32BEAV_CODEC_ID_PCM_F32LEAV_CODEC_ID_PCM_F64BEAV_CODEC_ID_PCM_F64LEAV_CODEC_ID_PCM_BLURAYAV_CODEC_ID_PCM_LXFAV_CODEC_ID_S302MAV_CODEC_ID_PCM_S8_PLANARAV_CODEC_ID_PCM_S24LE_PLANARAV_CODEC_ID_PCM_S32LE_PLANARAV_CODEC_ID_PCM_S16BE_PLANARAV_CODEC_ID_PCM_S64LEAV_CODEC_ID_PCM_S64BEAV_CODEC_ID_PCM_F16LEAV_CODEC_ID_PCM_F24LEAV_CODEC_ID_ADPCM_IMA_QTAV_CODEC_ID_ADPCM_IMA_WAVAV_CODEC_ID_ADPCM_IMA_DK3AV_CODEC_ID_ADPCM_IMA_DK4AV_CODEC_ID_ADPCM_IMA_WSAV_CODEC_ID_ADPCM_IMA_SMJPEGAV_CODEC_ID_ADPCM_MSAV_CODEC_ID_ADPCM_4XMAV_CODEC_ID_ADPCM_XAAV_CODEC_ID_ADPCM_ADXAV_CODEC_ID_ADPCM_EAAV_CODEC_ID_ADPCM_G726AV_CODEC_ID_ADPCM_CTAV_CODEC_ID_ADPCM_SWFAV_CODEC_ID_ADPCM_YAMAHAAV_CODEC_ID_ADPCM_SBPRO_4AV_CODEC_ID_ADPCM_SBPRO_3AV_CODEC_ID_ADPCM_SBPRO_2AV_CODEC_ID_ADPCM_THPAV_CODEC_ID_ADPCM_IMA_AMVAV_CODEC_ID_ADPCM_EA_R1AV_CODEC_ID_ADPCM_EA_R3AV_CODEC_ID_ADPCM_EA_R2AV_CODEC_ID_ADPCM_IMA_EA_SEADAV_CODEC_ID_ADPCM_IMA_EA_EACSAV_CODEC_ID_ADPCM_EA_XASAV_CODEC_ID_ADPCM_EA_MAXIS_XAAV_CODEC_ID_ADPCM_IMA_ISSAV_CODEC_ID_ADPCM_G722AV_CODEC_ID_ADPCM_IMA_APCAV_CODEC_ID_ADPCM_VIMAAV_CODEC_ID_ADPCM_AFCAV_CODEC_ID_ADPCM_IMA_OKIAV_CODEC_ID_ADPCM_DTKAV_CODEC_ID_ADPCM_IMA_RADAV_CODEC_ID_ADPCM_G726LEAV_CODEC_ID_ADPCM_THP_LEAV_CODEC_ID_ADPCM_PSXAV_CODEC_ID_ADPCM_AICAAV_CODEC_ID_ADPCM_IMA_DAT4AV_CODEC_ID_ADPCM_MTAFAV_CODEC_ID_AMR_NBAV_CODEC_ID_AMR_WBAV_CODEC_ID_RA_144AV_CODEC_ID_RA_288AV_CODEC_ID_ROQ_DPCMAV_CODEC_ID_INTERPLAY_DPCMAV_CODEC_ID_XAN_DPCMAV_CODEC_ID_SOL_DPCMAV_CODEC_ID_SDX2_DPCMAV_CODEC_ID_MP2AV_CODEC_ID_MP3AV_CODEC_ID_AACAV_CODEC_ID_AC3AV_CODEC_ID_DTSAV_CODEC_ID_VORBISAV_CODEC_ID_DVAUDIOAV_CODEC_ID_WMAV1AV_CODEC_ID_WMAV2AV_CODEC_ID_MACE3AV_CODEC_ID_MACE6AV_CODEC_ID_VMDAUDIOAV_CODEC_ID_FLACAV_CODEC_ID_MP3ADUAV_CODEC_ID_MP3ON4AV_CODEC_ID_SHORTENAV_CODEC_ID_ALACAV_CODEC_ID_WESTWOOD_SND1AV_CODEC_ID_GSMAV_CODEC_ID_QDM2AV_CODEC_ID_COOKAV_CODEC_ID_TRUESPEECHAV_CODEC_ID_TTAAV_CODEC_ID_SMACKAUDIOAV_CODEC_ID_QCELPAV_CODEC_ID_WAVPACKAV_CODEC_ID_DSICINAUDIOAV_CODEC_ID_IMCAV_CODEC_ID_MUSEPACK7AV_CODEC_ID_MLPAV_CODEC_ID_GSM_MSAV_CODEC_ID_ATRAC3AV_CODEC_ID_VOXWAREAV_CODEC_ID_APEAV_CODEC_ID_NELLYMOSERAV_CODEC_ID_MUSEPACK8AV_CODEC_ID_SPEEXAV_CODEC_ID_WMAVOICEAV_CODEC_ID_WMAPROAV_CODEC_ID_WMALOSSLESSAV_CODEC_ID_ATRAC3PAV_CODEC_ID_EAC3AV_CODEC_ID_SIPRAV_CODEC_ID_MP1AV_CODEC_ID_TWINVQAV_CODEC_ID_TRUEHDAV_CODEC_ID_MP4ALSAV_CODEC_ID_ATRAC1AV_CODEC_ID_BINKAUDIO_RDFTAV_CODEC_ID_BINKAUDIO_DCTAV_CODEC_ID_AAC_LATMAV_CODEC_ID_QDMCAV_CODEC_ID_CELTAV_CODEC_ID_G723_1AV_CODEC_ID_G729AV_CODEC_ID_8SVX_EXPAV_CODEC_ID_8SVX_FIBAV_CODEC_ID_BMV_AUDIOAV_CODEC_ID_RALFAV_CODEC_ID_IACAV_CODEC_ID_ILBCAV_CODEC_ID_OPUSAV_CODEC_ID_COMFORT_NOISEAV_CODEC_ID_TAKAV_CODEC_ID_METASOUNDAV_CODEC_ID_PAF_AUDIOAV_CODEC_ID_ON2AVCAV_CODEC_ID_DSS_SPAV_CODEC_ID_FFWAVESYNTHAV_CODEC_ID_SONICAV_CODEC_ID_SONIC_LSAV_CODEC_ID_EVRCAV_CODEC_ID_SMVAV_CODEC_ID_DSD_LSBFAV_CODEC_ID_DSD_MSBFAV_CODEC_ID_DSD_LSBF_PLANARAV_CODEC_ID_DSD_MSBF_PLANARAV_CODEC_ID_4GVAV_CODEC_ID_INTERPLAY_ACMAV_CODEC_ID_XMA1AV_CODEC_ID_XMA2AV_CODEC_ID_DSTAV_CODEC_ID_ATRAC3ALAV_CODEC_ID_ATRAC3PALAV_CODEC_ID_FIRST_SUBTITLEAV_CODEC_ID_DVB_SUBTITLEAV_CODEC_ID_TEXTAV_CODEC_ID_XSUBAV_CODEC_ID_SSAAV_CODEC_ID_MOV_TEXTAV_CODEC_ID_HDMV_PGS_SUBTITLEAV_CODEC_ID_DVB_TELETEXTAV_CODEC_ID_SRTAV_CODEC_ID_MICRODVDAV_CODEC_ID_EIA_608AV_CODEC_ID_JACOSUBAV_CODEC_ID_SAMIAV_CODEC_ID_REALTEXTAV_CODEC_ID_STLAV_CODEC_ID_SUBVIEWER1AV_CODEC_ID_SUBVIEWERAV_CODEC_ID_SUBRIPAV_CODEC_ID_WEBVTTAV_CODEC_ID_MPL2AV_CODEC_ID_VPLAYERAV_CODEC_ID_PJSAV_CODEC_ID_ASSAV_CODEC_ID_HDMV_TEXT_SUBTITLEAV_CODEC_ID_FIRST_UNKNOWNAV_CODEC_ID_SCTE_35AV_CODEC_ID_BINTEXTAV_CODEC_ID_XBINAV_CODEC_ID_IDFAV_CODEC_ID_OTFAV_CODEC_ID_SMPTE_KLVAV_CODEC_ID_DVD_NAVAV_CODEC_ID_TIMED_ID3AV_CODEC_ID_BIN_DATAAV_CODEC_ID_PROBEAV_CODEC_ID_MPEG2TSAV_CODEC_ID_MPEG4SYSTEMSAV_CODEC_ID_FFMETADATAAV_CODEC_ID_WRAPPED_AVFRAME