Enum av_sys::AVPixelFormat

#[repr(i32)]

pub enum AVPixelFormat {
    AV_PIX_FMT_NONE,
    AV_PIX_FMT_YUV420P,
    AV_PIX_FMT_YUYV422,
    AV_PIX_FMT_RGB24,
    AV_PIX_FMT_BGR24,
    AV_PIX_FMT_YUV422P,
    AV_PIX_FMT_YUV444P,
    AV_PIX_FMT_YUV410P,
    AV_PIX_FMT_YUV411P,
    AV_PIX_FMT_GRAY8,
    AV_PIX_FMT_MONOWHITE,
    AV_PIX_FMT_MONOBLACK,
    AV_PIX_FMT_PAL8,
    AV_PIX_FMT_YUVJ420P,
    AV_PIX_FMT_YUVJ422P,
    AV_PIX_FMT_YUVJ444P,
    AV_PIX_FMT_XVMC_MPEG2_MC,
    AV_PIX_FMT_XVMC_MPEG2_IDCT,
    AV_PIX_FMT_UYVY422,
    AV_PIX_FMT_UYYVYY411,
    AV_PIX_FMT_BGR8,
    AV_PIX_FMT_BGR4,
    AV_PIX_FMT_BGR4_BYTE,
    AV_PIX_FMT_RGB8,
    AV_PIX_FMT_RGB4,
    AV_PIX_FMT_RGB4_BYTE,
    AV_PIX_FMT_NV12,
    AV_PIX_FMT_NV21,
    AV_PIX_FMT_ARGB,
    AV_PIX_FMT_RGBA,
    AV_PIX_FMT_ABGR,
    AV_PIX_FMT_BGRA,
    AV_PIX_FMT_GRAY16BE,
    AV_PIX_FMT_GRAY16LE,
    AV_PIX_FMT_YUV440P,
    AV_PIX_FMT_YUVJ440P,
    AV_PIX_FMT_YUVA420P,
    AV_PIX_FMT_VDPAU_H264,
    AV_PIX_FMT_VDPAU_MPEG1,
    AV_PIX_FMT_VDPAU_MPEG2,
    AV_PIX_FMT_VDPAU_WMV3,
    AV_PIX_FMT_VDPAU_VC1,
    AV_PIX_FMT_RGB48BE,
    AV_PIX_FMT_RGB48LE,
    AV_PIX_FMT_RGB565BE,
    AV_PIX_FMT_RGB565LE,
    AV_PIX_FMT_RGB555BE,
    AV_PIX_FMT_RGB555LE,
    AV_PIX_FMT_BGR565BE,
    AV_PIX_FMT_BGR565LE,
    AV_PIX_FMT_BGR555BE,
    AV_PIX_FMT_BGR555LE,
    AV_PIX_FMT_VAAPI_MOCO,
    AV_PIX_FMT_VAAPI_IDCT,
    AV_PIX_FMT_VAAPI_VLD,
    AV_PIX_FMT_YUV420P16LE,
    AV_PIX_FMT_YUV420P16BE,
    AV_PIX_FMT_YUV422P16LE,
    AV_PIX_FMT_YUV422P16BE,
    AV_PIX_FMT_YUV444P16LE,
    AV_PIX_FMT_YUV444P16BE,
    AV_PIX_FMT_VDPAU_MPEG4,
    AV_PIX_FMT_DXVA2_VLD,
    AV_PIX_FMT_RGB444LE,
    AV_PIX_FMT_RGB444BE,
    AV_PIX_FMT_BGR444LE,
    AV_PIX_FMT_BGR444BE,
    AV_PIX_FMT_YA8,
    AV_PIX_FMT_BGR48BE,
    AV_PIX_FMT_BGR48LE,
    AV_PIX_FMT_YUV420P9BE,
    AV_PIX_FMT_YUV420P9LE,
    AV_PIX_FMT_YUV420P10BE,
    AV_PIX_FMT_YUV420P10LE,
    AV_PIX_FMT_YUV422P10BE,
    AV_PIX_FMT_YUV422P10LE,
    AV_PIX_FMT_YUV444P9BE,
    AV_PIX_FMT_YUV444P9LE,
    AV_PIX_FMT_YUV444P10BE,
    AV_PIX_FMT_YUV444P10LE,
    AV_PIX_FMT_YUV422P9BE,
    AV_PIX_FMT_YUV422P9LE,
    AV_PIX_FMT_VDA_VLD,
    AV_PIX_FMT_GBRP,
    AV_PIX_FMT_GBRP9BE,
    AV_PIX_FMT_GBRP9LE,
    AV_PIX_FMT_GBRP10BE,
    AV_PIX_FMT_GBRP10LE,
    AV_PIX_FMT_GBRP16BE,
    AV_PIX_FMT_GBRP16LE,
    AV_PIX_FMT_YUVA422P,
    AV_PIX_FMT_YUVA444P,
    AV_PIX_FMT_YUVA420P9BE,
    AV_PIX_FMT_YUVA420P9LE,
    AV_PIX_FMT_YUVA422P9BE,
    AV_PIX_FMT_YUVA422P9LE,
    AV_PIX_FMT_YUVA444P9BE,
    AV_PIX_FMT_YUVA444P9LE,
    AV_PIX_FMT_YUVA420P10BE,
    AV_PIX_FMT_YUVA420P10LE,
    AV_PIX_FMT_YUVA422P10BE,
    AV_PIX_FMT_YUVA422P10LE,
    AV_PIX_FMT_YUVA444P10BE,
    AV_PIX_FMT_YUVA444P10LE,
    AV_PIX_FMT_YUVA420P16BE,
    AV_PIX_FMT_YUVA420P16LE,
    AV_PIX_FMT_YUVA422P16BE,
    AV_PIX_FMT_YUVA422P16LE,
    AV_PIX_FMT_YUVA444P16BE,
    AV_PIX_FMT_YUVA444P16LE,
    AV_PIX_FMT_VDPAU,
    AV_PIX_FMT_XYZ12LE,
    AV_PIX_FMT_XYZ12BE,
    AV_PIX_FMT_NV16,
    AV_PIX_FMT_NV20LE,
    AV_PIX_FMT_NV20BE,
    AV_PIX_FMT_RGBA64BE,
    AV_PIX_FMT_RGBA64LE,
    AV_PIX_FMT_BGRA64BE,
    AV_PIX_FMT_BGRA64LE,
    AV_PIX_FMT_YVYU422,
    AV_PIX_FMT_VDA,
    AV_PIX_FMT_YA16BE,
    AV_PIX_FMT_YA16LE,
    AV_PIX_FMT_GBRAP,
    AV_PIX_FMT_GBRAP16BE,
    AV_PIX_FMT_GBRAP16LE,
    AV_PIX_FMT_QSV,
    AV_PIX_FMT_MMAL,
    AV_PIX_FMT_D3D11VA_VLD,
    AV_PIX_FMT_CUDA,
    AV_PIX_FMT_0RGB,
    AV_PIX_FMT_RGB0,
    AV_PIX_FMT_0BGR,
    AV_PIX_FMT_BGR0,
    AV_PIX_FMT_YUV420P12BE,
    AV_PIX_FMT_YUV420P12LE,
    AV_PIX_FMT_YUV420P14BE,
    AV_PIX_FMT_YUV420P14LE,
    AV_PIX_FMT_YUV422P12BE,
    AV_PIX_FMT_YUV422P12LE,
    AV_PIX_FMT_YUV422P14BE,
    AV_PIX_FMT_YUV422P14LE,
    AV_PIX_FMT_YUV444P12BE,
    AV_PIX_FMT_YUV444P12LE,
    AV_PIX_FMT_YUV444P14BE,
    AV_PIX_FMT_YUV444P14LE,
    AV_PIX_FMT_GBRP12BE,
    AV_PIX_FMT_GBRP12LE,
    AV_PIX_FMT_GBRP14BE,
    AV_PIX_FMT_GBRP14LE,
    AV_PIX_FMT_YUVJ411P,
    AV_PIX_FMT_BAYER_BGGR8,
    AV_PIX_FMT_BAYER_RGGB8,
    AV_PIX_FMT_BAYER_GBRG8,
    AV_PIX_FMT_BAYER_GRBG8,
    AV_PIX_FMT_BAYER_BGGR16LE,
    AV_PIX_FMT_BAYER_BGGR16BE,
    AV_PIX_FMT_BAYER_RGGB16LE,
    AV_PIX_FMT_BAYER_RGGB16BE,
    AV_PIX_FMT_BAYER_GBRG16LE,
    AV_PIX_FMT_BAYER_GBRG16BE,
    AV_PIX_FMT_BAYER_GRBG16LE,
    AV_PIX_FMT_BAYER_GRBG16BE,
    AV_PIX_FMT_YUV440P10LE,
    AV_PIX_FMT_YUV440P10BE,
    AV_PIX_FMT_YUV440P12LE,
    AV_PIX_FMT_YUV440P12BE,
    AV_PIX_FMT_AYUV64LE,
    AV_PIX_FMT_AYUV64BE,
    AV_PIX_FMT_VIDEOTOOLBOX,
    AV_PIX_FMT_P010LE,
    AV_PIX_FMT_P010BE,
    AV_PIX_FMT_GBRAP12BE,
    AV_PIX_FMT_GBRAP12LE,
    AV_PIX_FMT_GBRAP10BE,
    AV_PIX_FMT_GBRAP10LE,
    AV_PIX_FMT_MEDIACODEC,
    AV_PIX_FMT_GRAY12BE,
    AV_PIX_FMT_GRAY12LE,
    AV_PIX_FMT_GRAY10BE,
    AV_PIX_FMT_GRAY10LE,
    AV_PIX_FMT_P016LE,
    AV_PIX_FMT_P016BE,
    AV_PIX_FMT_NB,
}

Pixel format.

@note AV_PIX_FMT_RGB32 is handled in an endian-specific manner. An RGBA color is put together as: (A << 24) | (R << 16) | (G << 8) | B This is stored as BGRA on little-endian CPU architectures and ARGB on big-endian CPUs.

@par When the pixel format is palettized RGB32 (AV_PIX_FMT_PAL8), the palettized image data is stored in AVFrame.data[0]. The palette is transported in AVFrame.data[1], is 1024 bytes long (256 4-byte entries) and is formatted the same as in AV_PIX_FMT_RGB32 described above (i.e., it is also endian-specific). Note also that the individual RGB32 palette components stored in AVFrame.data[1] should be in the range 0..255. This is important as many custom PAL8 video codecs that were designed to run on the IBM VGA graphics adapter use 6-bit palette components.

@par For all the 8 bits per pixel formats, an RGB32 palette is in data[1] like for pal8. This palette is filled in automatically by the function allocating the picture.

Variants
AV_PIX_FMT_NONEAV_PIX_FMT_YUV420PAV_PIX_FMT_YUYV422AV_PIX_FMT_RGB24AV_PIX_FMT_BGR24AV_PIX_FMT_YUV422PAV_PIX_FMT_YUV444PAV_PIX_FMT_YUV410PAV_PIX_FMT_YUV411PAV_PIX_FMT_GRAY8AV_PIX_FMT_MONOWHITEAV_PIX_FMT_MONOBLACKAV_PIX_FMT_PAL8AV_PIX_FMT_YUVJ420PAV_PIX_FMT_YUVJ422PAV_PIX_FMT_YUVJ444PAV_PIX_FMT_XVMC_MPEG2_MCAV_PIX_FMT_XVMC_MPEG2_IDCTAV_PIX_FMT_UYVY422AV_PIX_FMT_UYYVYY411AV_PIX_FMT_BGR8AV_PIX_FMT_BGR4AV_PIX_FMT_BGR4_BYTEAV_PIX_FMT_RGB8AV_PIX_FMT_RGB4AV_PIX_FMT_RGB4_BYTEAV_PIX_FMT_NV12AV_PIX_FMT_NV21AV_PIX_FMT_ARGBAV_PIX_FMT_RGBAAV_PIX_FMT_ABGRAV_PIX_FMT_BGRAAV_PIX_FMT_GRAY16BEAV_PIX_FMT_GRAY16LEAV_PIX_FMT_YUV440PAV_PIX_FMT_YUVJ440PAV_PIX_FMT_YUVA420PAV_PIX_FMT_VDPAU_H264AV_PIX_FMT_VDPAU_MPEG1AV_PIX_FMT_VDPAU_MPEG2AV_PIX_FMT_VDPAU_WMV3AV_PIX_FMT_VDPAU_VC1AV_PIX_FMT_RGB48BEAV_PIX_FMT_RGB48LEAV_PIX_FMT_RGB565BEAV_PIX_FMT_RGB565LEAV_PIX_FMT_RGB555BEAV_PIX_FMT_RGB555LEAV_PIX_FMT_BGR565BEAV_PIX_FMT_BGR565LEAV_PIX_FMT_BGR555BEAV_PIX_FMT_BGR555LEAV_PIX_FMT_VAAPI_MOCOAV_PIX_FMT_VAAPI_IDCTAV_PIX_FMT_VAAPI_VLDAV_PIX_FMT_YUV420P16LEAV_PIX_FMT_YUV420P16BEAV_PIX_FMT_YUV422P16LEAV_PIX_FMT_YUV422P16BEAV_PIX_FMT_YUV444P16LEAV_PIX_FMT_YUV444P16BEAV_PIX_FMT_VDPAU_MPEG4AV_PIX_FMT_DXVA2_VLDAV_PIX_FMT_RGB444LEAV_PIX_FMT_RGB444BEAV_PIX_FMT_BGR444LEAV_PIX_FMT_BGR444BEAV_PIX_FMT_YA8AV_PIX_FMT_BGR48BEAV_PIX_FMT_BGR48LEAV_PIX_FMT_YUV420P9BEAV_PIX_FMT_YUV420P9LEAV_PIX_FMT_YUV420P10BEAV_PIX_FMT_YUV420P10LEAV_PIX_FMT_YUV422P10BEAV_PIX_FMT_YUV422P10LEAV_PIX_FMT_YUV444P9BEAV_PIX_FMT_YUV444P9LEAV_PIX_FMT_YUV444P10BEAV_PIX_FMT_YUV444P10LEAV_PIX_FMT_YUV422P9BEAV_PIX_FMT_YUV422P9LEAV_PIX_FMT_VDA_VLDAV_PIX_FMT_GBRPAV_PIX_FMT_GBRP9BEAV_PIX_FMT_GBRP9LEAV_PIX_FMT_GBRP10BEAV_PIX_FMT_GBRP10LEAV_PIX_FMT_GBRP16BEAV_PIX_FMT_GBRP16LEAV_PIX_FMT_YUVA422PAV_PIX_FMT_YUVA444PAV_PIX_FMT_YUVA420P9BEAV_PIX_FMT_YUVA420P9LEAV_PIX_FMT_YUVA422P9BEAV_PIX_FMT_YUVA422P9LEAV_PIX_FMT_YUVA444P9BEAV_PIX_FMT_YUVA444P9LEAV_PIX_FMT_YUVA420P10BEAV_PIX_FMT_YUVA420P10LEAV_PIX_FMT_YUVA422P10BEAV_PIX_FMT_YUVA422P10LEAV_PIX_FMT_YUVA444P10BEAV_PIX_FMT_YUVA444P10LEAV_PIX_FMT_YUVA420P16BEAV_PIX_FMT_YUVA420P16LEAV_PIX_FMT_YUVA422P16BEAV_PIX_FMT_YUVA422P16LEAV_PIX_FMT_YUVA444P16BEAV_PIX_FMT_YUVA444P16LEAV_PIX_FMT_VDPAUAV_PIX_FMT_XYZ12LEAV_PIX_FMT_XYZ12BEAV_PIX_FMT_NV16AV_PIX_FMT_NV20LEAV_PIX_FMT_NV20BEAV_PIX_FMT_RGBA64BEAV_PIX_FMT_RGBA64LEAV_PIX_FMT_BGRA64BEAV_PIX_FMT_BGRA64LEAV_PIX_FMT_YVYU422AV_PIX_FMT_VDAAV_PIX_FMT_YA16BEAV_PIX_FMT_YA16LEAV_PIX_FMT_GBRAPAV_PIX_FMT_GBRAP16BEAV_PIX_FMT_GBRAP16LEAV_PIX_FMT_QSVAV_PIX_FMT_MMALAV_PIX_FMT_D3D11VA_VLDAV_PIX_FMT_CUDAAV_PIX_FMT_0RGBAV_PIX_FMT_RGB0AV_PIX_FMT_0BGRAV_PIX_FMT_BGR0AV_PIX_FMT_YUV420P12BEAV_PIX_FMT_YUV420P12LEAV_PIX_FMT_YUV420P14BEAV_PIX_FMT_YUV420P14LEAV_PIX_FMT_YUV422P12BEAV_PIX_FMT_YUV422P12LEAV_PIX_FMT_YUV422P14BEAV_PIX_FMT_YUV422P14LEAV_PIX_FMT_YUV444P12BEAV_PIX_FMT_YUV444P12LEAV_PIX_FMT_YUV444P14BEAV_PIX_FMT_YUV444P14LEAV_PIX_FMT_GBRP12BEAV_PIX_FMT_GBRP12LEAV_PIX_FMT_GBRP14BEAV_PIX_FMT_GBRP14LEAV_PIX_FMT_YUVJ411PAV_PIX_FMT_BAYER_BGGR8AV_PIX_FMT_BAYER_RGGB8AV_PIX_FMT_BAYER_GBRG8AV_PIX_FMT_BAYER_GRBG8AV_PIX_FMT_BAYER_BGGR16LEAV_PIX_FMT_BAYER_BGGR16BEAV_PIX_FMT_BAYER_RGGB16LEAV_PIX_FMT_BAYER_RGGB16BEAV_PIX_FMT_BAYER_GBRG16LEAV_PIX_FMT_BAYER_GBRG16BEAV_PIX_FMT_BAYER_GRBG16LEAV_PIX_FMT_BAYER_GRBG16BEAV_PIX_FMT_YUV440P10LEAV_PIX_FMT_YUV440P10BEAV_PIX_FMT_YUV440P12LEAV_PIX_FMT_YUV440P12BEAV_PIX_FMT_AYUV64LEAV_PIX_FMT_AYUV64BEAV_PIX_FMT_VIDEOTOOLBOXAV_PIX_FMT_P010LEAV_PIX_FMT_P010BEAV_PIX_FMT_GBRAP12BEAV_PIX_FMT_GBRAP12LEAV_PIX_FMT_GBRAP10BEAV_PIX_FMT_GBRAP10LEAV_PIX_FMT_MEDIACODECAV_PIX_FMT_GRAY12BEAV_PIX_FMT_GRAY12LEAV_PIX_FMT_GRAY10BEAV_PIX_FMT_GRAY10LEAV_PIX_FMT_P016LEAV_PIX_FMT_P016BEAV_PIX_FMT_NB