Function av_sys::av_audio_resample_init

pub unsafe extern fn av_audio_resample_init(
    output_channels: c_int, 
    input_channels: c_int, 
    output_rate: c_int, 
    input_rate: c_int, 
    sample_fmt_out: AVSampleFormat, 
    sample_fmt_in: AVSampleFormat, 
    filter_length: c_int, 
    log2_phase_count: c_int, 
    linear: c_int, 
    cutoff: f64
) -> *mut ReSampleContext

Initialize audio resampling context. *

@param output_channels number of output channels

@param input_channels number of input channels

@param output_rate output sample rate

@param input_rate input sample rate

@param sample_fmt_out requested output sample format

@param sample_fmt_in input sample format

@param filter_length length of each FIR filter in the filterbank relative to the cutoff frequency

@param log2_phase_count log2 of the number of entries in the polyphase filterbank

@param linear if 1 then the used FIR filter will be linearly interpolated between the 2 closest, if 0 the closest will be used

@param cutoff cutoff frequency, 1.0 corresponds to half the output sampling rate

@return allocated ReSampleContext, NULL if error occurred