Function av_sys::av_buffer_create

pub unsafe extern fn av_buffer_create(
    data: *mut u8, 
    size: c_int, 
    free: Option<unsafe extern fn(_: *mut c_void, _: *mut u8)>, 
    opaque: *mut c_void, 
    flags: c_int
) -> *mut AVBufferRef

Create an AVBuffer from an existing array.

If this function is successful, data is owned by the AVBuffer. The caller may only access data through the returned AVBufferRef and references derived from it. If this function fails, data is left untouched. @param data data array @param size size of data in bytes @param free a callback for freeing this buffer's data @param opaque parameter to be got for processing or passed to free @param flags a combination of AV_BUFFER_FLAG_*

@return an AVBufferRef referring to data on success, NULL on failure.