Function av_sys::av_image_check_size2

pub unsafe extern fn av_image_check_size2(
    w: c_uint, 
    h: c_uint, 
    max_pixels: i64, 
    pix_fmt: AVPixelFormat, 
    log_offset: c_int, 
    log_ctx: *mut c_void
) -> c_int

Check if the given dimension of an image is valid, meaning that all bytes of a plane of an image with the specified pix_fmt can be addressed with a signed int.

@param w the width of the picture @param h the height of the picture @param max_pixels the maximum number of pixels the user wants to accept @param pix_fmt the pixel format, can be AV_PIX_FMT_NONE if unknown. @param log_offset the offset to sum to the log level for logging with log_ctx @param log_ctx the parent logging context, it may be NULL @return >= 0 if valid, a negative error code otherwise