Function av_sys::av_interleaved_write_uncoded_frame

pub unsafe extern fn av_interleaved_write_uncoded_frame(
    s: *mut AVFormatContext, 
    stream_index: c_int, 
    frame: *mut AVFrame
) -> c_int

Write an uncoded frame to an output media file.

If the muxer supports it, this function makes it possible to write an AVFrame structure directly, without encoding it into a packet. It is mostly useful for devices and similar special muxers that use raw video or PCM data and will not serialize it into a byte stream.

To test whether it is possible to use it with a given muxer and stream, use av_write_uncoded_frame_query().

The caller gives up ownership of the frame and must not access it afterwards.

@return >=0 for success, a negative code on error