Function av_sys::av_log_default_callback

pub unsafe extern fn av_log_default_callback(
    avcl: *mut c_void, 
    level: c_int, 
    fmt: *const c_char, 
    vl: *mut __va_list_tag
)

Default logging callback

It prints the message to stderr, optionally colorizing it.

@param avcl A pointer to an arbitrary struct of which the first field is a pointer to an AVClass struct. @param level The importance level of the message expressed using a @ref lavu_log_constants "Logging Constant". @param fmt The format string (printf-compatible) that specifies how subsequent arguments are converted to output. @param vl The arguments referenced by the format string.