Function av_sys::av_memcpy_backptr

pub unsafe extern fn av_memcpy_backptr(dst: *mut u8, back: c_int, cnt: c_int)

Overlapping memcpy() implementation.

@param dst Destination buffer @param back Number of bytes back to start copying (i.e. the initial size of the overlapping window); must be > 0 @param cnt Number of bytes to copy; must be >= 0

@note cnt > back is valid, this will copy the bytes we just copied, thus creating a repeating pattern with a period length of back.