Function av_sys::av_probe_input_buffer2

pub unsafe extern fn av_probe_input_buffer2(
    pb: *mut AVIOContext, 
    fmt: *mut *mut AVInputFormat, 
    url: *const c_char, 
    logctx: *mut c_void, 
    offset: c_uint, 
    max_probe_size: c_uint
) -> c_int

Probe a bytestream to determine the input format. Each time a probe returns with a score that is too low, the probe buffer size is increased and another attempt is made. When the maximum probe size is reached, the input format with the highest score is returned.

@param pb the bytestream to probe @param fmt the input format is put here @param url the url of the stream @param logctx the log context @param offset the offset within the bytestream to probe from @param max_probe_size the maximum probe buffer size (zero for default) @return the score in case of success, a negative value corresponding to an the maximal score is AVPROBE_SCORE_MAX AVERROR code otherwise