Function av_sys::av_read_image_line

pub unsafe extern fn av_read_image_line(
    dst: *mut u16, 
    data: *mut *const u8, 
    linesize: *const c_int, 
    desc: *const AVPixFmtDescriptor, 
    x: c_int, 
    y: c_int, 
    c: c_int, 
    w: c_int, 
    read_pal_component: c_int
)

Read a line from an image, and write the values of the pixel format component c to dst.

@param data the array containing the pointers to the planes of the image @param linesize the array containing the linesizes of the image @param desc the pixel format descriptor for the image @param x the horizontal coordinate of the first pixel to read @param y the vertical coordinate of the first pixel to read @param w the width of the line to read, that is the number of values to write to dst @param read_pal_component if not zero and the format is a paletted format writes the values corresponding to the palette component c in data[1] to dst, rather than the palette indexes in data[0]. The behavior is undefined if the format is not paletted.