Function av_sys::av_realloc

pub unsafe extern fn av_realloc(ptr: *mut c_void, size: usize) -> *mut c_void

Allocate, reallocate, or free a block of memory.

If ptr is NULL and size > 0, allocate a new block. If size is zero, free the memory block pointed to by ptr. Otherwise, expand or shrink that block of memory according to size.

@param ptr Pointer to a memory block already allocated with av_realloc() or NULL @param size Size in bytes of the memory block to be allocated or reallocated

@return Pointer to a newly-reallocated block or NULL if the block cannot be reallocated or the function is used to free the memory block

@warning Unlike av_malloc(), the returned pointer is not guaranteed to be correctly aligned. @see av_fast_realloc() @see av_reallocp()