Function av_sys::av_realloc_array

pub unsafe extern fn av_realloc_array(
    ptr: *mut c_void, 
    nmemb: usize, 
    size: usize
) -> *mut c_void

Allocate, reallocate, or free an array.

If ptr is NULL and nmemb > 0, allocate a new block. If nmemb is zero, free the memory block pointed to by ptr.

@param ptr Pointer to a memory block already allocated with av_realloc() or NULL @param nmemb Number of elements in the array @param size Size of the single element of the array

@return Pointer to a newly-reallocated block or NULL if the block cannot be reallocated or the function is used to free the memory block

@warning Unlike av_malloc(), the allocated memory is not guaranteed to be correctly aligned. @see av_reallocp_array()