Function av_sys::av_url_split

pub unsafe extern fn av_url_split(
    proto: *mut c_char, 
    proto_size: c_int, 
    authorization: *mut c_char, 
    authorization_size: c_int, 
    hostname: *mut c_char, 
    hostname_size: c_int, 
    port_ptr: *mut c_int, 
    path: *mut c_char, 
    path_size: c_int, 
    url: *const c_char
)

Split a URL string into components.

The pointers to buffers for storing individual components may be null, in order to ignore that component. Buffers for components not found are set to empty strings. If the port is not found, it is set to a negative value.

@param proto the buffer for the protocol @param proto_size the size of the proto buffer @param authorization the buffer for the authorization @param authorization_size the size of the authorization buffer @param hostname the buffer for the host name @param hostname_size the size of the hostname buffer @param port_ptr a pointer to store the port number in @param path the buffer for the path @param path_size the size of the path buffer @param url the URL to split