Function av_sys::avcodec_decode_audio4

pub unsafe extern fn avcodec_decode_audio4(
    avctx: *mut AVCodecContext, 
    frame: *mut AVFrame, 
    got_frame_ptr: *mut c_int, 
    avpkt: *const AVPacket
) -> c_int

Decode the audio frame of size avpkt->size from avpkt->data into frame. *

Some decoders may support multiple frames in a single AVPacket. Such

decoders would then just decode the first frame and the return value would be

less than the packet size. In this case, avcodec_decode_audio4 has to be

called again with an AVPacket containing the remaining data in order to

decode the second frame, etc... Even if no frames are returned, the packet

needs to be fed to the decoder with remaining data until it is completely

consumed or an error occurs. *

Some decoders (those marked with AV_CODEC_CAP_DELAY) have a delay between input

and output. This means that for some packets they will not immediately

produce decoded output and need to be flushed at the end of decoding to get

all the decoded data. Flushing is done by calling this function with packets

with avpkt->data set to NULL and avpkt->size set to 0 until it stops

returning samples. It is safe to flush even those decoders that are not

marked with AV_CODEC_CAP_DELAY, then no samples will be returned. *

@warning The input buffer, avpkt->data must be AV_INPUT_BUFFER_PADDING_SIZE

larger than the actual read bytes because some optimized bitstream

readers read 32 or 64 bits at once and could read over the end. *

@note The AVCodecContext MUST have been opened with @ref avcodec_open2()

before packets may be fed to the decoder. *

@param avctx the codec context

@param[out] frame The AVFrame in which to store decoded audio samples.

The decoder will allocate a buffer for the decoded frame by

calling the AVCodecContext.get_buffer2() callback.

When AVCodecContext.refcounted_frames is set to 1, the frame is

reference counted and the returned reference belongs to the

caller. The caller must release the frame using av_frame_unref()

when the frame is no longer needed. The caller may safely write

to the frame if av_frame_is_writable() returns 1.

When AVCodecContext.refcounted_frames is set to 0, the returned

reference belongs to the decoder and is valid only until the

next call to this function or until closing or flushing the

decoder. The caller may not write to it.

@param[out] got_frame_ptr Zero if no frame could be decoded, otherwise it is

non-zero. Note that this field being set to zero

does not mean that an error has occurred. For

decoders with AV_CODEC_CAP_DELAY set, no given decode

call is guaranteed to produce a frame.

@param[in] avpkt The input AVPacket containing the input buffer.

At least avpkt->data and avpkt->size should be set. Some

decoders might also require additional fields to be set.

@return A negative error code is returned if an error occurred during

decoding, otherwise the number of bytes consumed from the input

AVPacket is returned. *

@deprecated Use avcodec_send_packet() and avcodec_receive_frame().