Function av_sys::avcodec_encode_video2

pub unsafe extern fn avcodec_encode_video2(
    avctx: *mut AVCodecContext, 
    avpkt: *mut AVPacket, 
    frame: *const AVFrame, 
    got_packet_ptr: *mut c_int
) -> c_int

Encode a frame of video.

Takes input raw video data from frame and writes the next output packet, if available, to avpkt. The output packet does not necessarily contain data for the most recent frame, as encoders can delay and reorder input frames internally as needed.

@param avctx codec context @param avpkt output AVPacket. The user can supply an output buffer by setting avpkt->data and avpkt->size prior to calling the function, but if the size of the user-provided data is not large enough, encoding will fail. All other AVPacket fields will be reset by the encoder using av_init_packet(). If avpkt->data is NULL, the encoder will allocate it. The encoder will set avpkt->size to the size of the output packet. The returned data (if any) belongs to the caller, he is responsible for freeing it.

If this function fails or produces no output, avpkt will be
             freed using av_packet_unref().

@param[in] frame AVFrame containing the raw video data to be encoded. May be NULL when flushing an encoder that has the AV_CODEC_CAP_DELAY capability set. @param[out] got_packet_ptr This field is set to 1 by libavcodec if the output packet is non-empty, and to 0 if it is empty. If the function returns an error, the packet can be assumed to be invalid, and the value of got_packet_ptr is undefined and should not be used. @return 0 on success, negative error code on failure

@deprecated use avcodec_send_frame()/avcodec_receive_packet() instead