Function av_sys::avformat_open_input

pub unsafe extern fn avformat_open_input(
    ps: *mut *mut AVFormatContext, 
    url: *const c_char, 
    fmt: *mut AVInputFormat, 
    options: *mut *mut AVDictionary
) -> c_int

Open an input stream and read the header. The codecs are not opened. The stream must be closed with avformat_close_input().

@param ps Pointer to user-supplied AVFormatContext (allocated by avformat_alloc_context). May be a pointer to NULL, in which case an AVFormatContext is allocated by this function and written into ps. Note that a user-supplied AVFormatContext will be freed on failure. @param url URL of the stream to open. @param fmt If non-NULL, this parameter forces a specific input format. Otherwise the format is autodetected. @param options A dictionary filled with AVFormatContext and demuxer-private options. On return this parameter will be destroyed and replaced with a dict containing options that were not found. May be NULL.

@return 0 on success, a negative AVERROR on failure.

@note If you want to use custom IO, preallocate the format context and set its pb field.