Function av_sys::avformat_seek_file

pub unsafe extern fn avformat_seek_file(
    s: *mut AVFormatContext, 
    stream_index: c_int, 
    min_ts: i64, 
    ts: i64, 
    max_ts: i64, 
    flags: c_int
) -> c_int

Seek to timestamp ts. Seeking will be done so that the point from which all active streams can be presented successfully will be closest to ts and within min/max_ts. Active streams are all streams that have AVStream.discard < AVDISCARD_ALL.

If flags contain AVSEEK_FLAG_BYTE, then all timestamps are in bytes and are the file position (this may not be supported by all demuxers). If flags contain AVSEEK_FLAG_FRAME, then all timestamps are in frames in the stream with stream_index (this may not be supported by all demuxers). Otherwise all timestamps are in units of the stream selected by stream_index or if stream_index is -1, in AV_TIME_BASE units. If flags contain AVSEEK_FLAG_ANY, then non-keyframes are treated as keyframes (this may not be supported by all demuxers). If flags contain AVSEEK_FLAG_BACKWARD, it is ignored.

@param s media file handle @param stream_index index of the stream which is used as time base reference @param min_ts smallest acceptable timestamp @param ts target timestamp @param max_ts largest acceptable timestamp @param flags flags @return >=0 on success, error code otherwise

@note This is part of the new seek API which is still under construction. Thus do not use this yet. It may change at any time, do not expect ABI compatibility yet!