Function av_sys::avio_seek_time

pub unsafe extern fn avio_seek_time(
    h: *mut AVIOContext, 
    stream_index: c_int, 
    timestamp: i64, 
    flags: c_int
) -> i64

Seek to a given timestamp relative to some component stream. Only meaningful if using a network streaming protocol (e.g. MMS.).

@param h IO context from which to call the seek function pointers @param stream_index The stream index that the timestamp is relative to. If stream_index is (-1) the timestamp should be in AV_TIME_BASE units from the beginning of the presentation. If a stream_index >= 0 is used and the protocol does not support seeking based on component streams, the call will fail. @param timestamp timestamp in AVStream.time_base units or if there is no stream specified then in AV_TIME_BASE units. @param flags Optional combination of AVSEEK_FLAG_BACKWARD, AVSEEK_FLAG_BYTE and AVSEEK_FLAG_ANY. The protocol may silently ignore AVSEEK_FLAG_BACKWARD and AVSEEK_FLAG_ANY, but AVSEEK_FLAG_BYTE will fail if used and not supported. @return >= 0 on success @see AVInputFormat::read_seek