Function av_sys::swr_get_out_samples

pub unsafe extern fn swr_get_out_samples(
    s: *mut SwrContext, 
    in_samples: c_int
) -> c_int

Find an upper bound on the number of samples that the next swr_convert call will output, if called with in_samples of input samples. This depends on the internal state, and anything changing the internal state (like further swr_convert() calls) will may change the number of samples swr_get_out_samples() returns for the same number of input samples.

@param in_samples number of input samples. @note any call to swr_inject_silence(), swr_convert(), swr_next_pts() or swr_set_compensation() invalidates this limit @note it is recommended to pass the correct available buffer size to all functions like swr_convert() even if swr_get_out_samples() indicates that less would be used. @returns an upper bound on the number of samples that the next swr_convert will output or a negative value to indicate an error