Function av_sys::sws_scale

pub unsafe extern fn sws_scale(
    c: *mut SwsContext, 
    srcSlice: *const *const u8, 
    srcStride: *const c_int, 
    srcSliceY: c_int, 
    srcSliceH: c_int, 
    dst: *const *const u8, 
    dstStride: *const c_int
) -> c_int

Scale the image slice in srcSlice and put the resulting scaled slice in the image in dst. A slice is a sequence of consecutive rows in an image.

Slices have to be provided in sequential order, either in top-bottom or bottom-top order. If slices are provided in non-sequential order the behavior of the function is undefined.

@param c the scaling context previously created with sws_getContext() @param srcSlice the array containing the pointers to the planes of the source slice @param srcStride the array containing the strides for each plane of the source image @param srcSliceY the position in the source image of the slice to process, that is the number (counted starting from zero) in the image of the first row of the slice @param srcSliceH the height of the source slice, that is the number of rows in the slice @param dst the array containing the pointers to the planes of the destination image @param dstStride the array containing the strides for each plane of the destination image @return the height of the output slice