Struct av_sys::AVBitStreamFilter

#[repr(C)]

pub struct AVBitStreamFilter {
    pub name: *const c_char,
    pub codec_ids: *const AVCodecID,
    pub priv_class: *const AVClass,
    pub priv_data_size: c_int,
    pub init: Option<unsafe extern fn(_: *mut AVBSFContext) -> c_int>,
    pub filter: Option<unsafe extern fn(_: *mut AVBSFContext, _: *mut AVPacket) -> c_int>,
    pub close: Option<unsafe extern fn(_: *mut AVBSFContext)>,
}

Fields
name: *const c_char codec_ids: *const AVCodecID

A list of codec ids supported by the filter, terminated by AV_CODEC_ID_NONE. May be NULL, in that case the bitstream filter works with any codec id.

priv_class: *const AVClass

A class for the private data, used to declare bitstream filter private AVOptions. This field is NULL for bitstream filters that do not declare any options.

If this field is non-NULL, the first member of the filter private data must be a pointer to AVClass, which will be set by libavcodec generic code to this class.

priv_data_size: c_int init: Option<unsafe extern fn(_: *mut AVBSFContext) -> c_int> filter: Option<unsafe extern fn(_: *mut AVBSFContext, _: *mut AVPacket) -> c_int> close: Option<unsafe extern fn(_: *mut AVBSFContext)>