Struct av_sys::AVCodec

#[repr(C)]

pub struct AVCodec {
    pub name: *const c_char,
    pub long_name: *const c_char,
    pub type_: AVMediaType,
    pub id: AVCodecID,
    pub capabilities: c_int,
    pub supported_framerates: *const AVRational,
    pub pix_fmts: *const AVPixelFormat,
    pub supported_samplerates: *const c_int,
    pub sample_fmts: *const AVSampleFormat,
    pub channel_layouts: *const u64,
    pub max_lowres: u8,
    pub priv_class: *const AVClass,
    pub profiles: *const AVProfile,
    pub priv_data_size: c_int,
    pub next: *mut AVCodec,
    pub init_thread_copy: Option<unsafe extern fn(_: *mut AVCodecContext) -> c_int>,
    pub update_thread_context: Option<unsafe extern fn(_: *mut AVCodecContext, _: *const AVCodecContext) -> c_int>,
    pub defaults: *const AVCodecDefault,
    pub init_static_data: Option<unsafe extern fn(_: *mut AVCodec)>,
    pub init: Option<unsafe extern fn(_: *mut AVCodecContext) -> c_int>,
    pub encode_sub: Option<unsafe extern fn(_: *mut AVCodecContext, _: *mut u8, _: c_int, _: *const AVSubtitle) -> c_int>,
    pub encode2: Option<unsafe extern fn(_: *mut AVCodecContext, _: *mut AVPacket, _: *const AVFrame, _: *mut c_int) -> c_int>,
    pub decode: Option<unsafe extern fn(_: *mut AVCodecContext, _: *mut c_void, _: *mut c_int, _: *mut AVPacket) -> c_int>,
    pub close: Option<unsafe extern fn(_: *mut AVCodecContext) -> c_int>,
    pub send_frame: Option<unsafe extern fn(_: *mut AVCodecContext, _: *const AVFrame) -> c_int>,
    pub send_packet: Option<unsafe extern fn(_: *mut AVCodecContext, _: *const AVPacket) -> c_int>,
    pub receive_frame: Option<unsafe extern fn(_: *mut AVCodecContext, _: *mut AVFrame) -> c_int>,
    pub receive_packet: Option<unsafe extern fn(_: *mut AVCodecContext, _: *mut AVPacket) -> c_int>,
    pub flush: Option<unsafe extern fn(_: *mut AVCodecContext)>,
    pub caps_internal: c_int,
}

AVCodec.

Fields
name: *const c_char

Name of the codec implementation. The name is globally unique among encoders and among decoders (but an encoder and a decoder can share the same name). This is the primary way to find a codec from the user perspective.

long_name: *const c_char

Descriptive name for the codec, meant to be more human readable than name. You should use the NULL_IF_CONFIG_SMALL() macro to define it.

type_: AVMediaType id: AVCodecID capabilities: c_int

Codec capabilities. see AV_CODEC_CAP_*

supported_framerates: *const AVRational

< array of supported framerates, or NULL if any, array is terminated by {0,0}

pix_fmts: *const AVPixelFormat

< array of supported pixel formats, or NULL if unknown, array is terminated by -1

supported_samplerates: *const c_int

< array of supported audio samplerates, or NULL if unknown, array is terminated by 0

sample_fmts: *const AVSampleFormat

< array of supported sample formats, or NULL if unknown, array is terminated by -1

channel_layouts: *const u64

< array of support channel layouts, or NULL if unknown. array is terminated by 0

max_lowres: u8

< maximum value for lowres supported by the decoder

priv_class: *const AVClass

< AVClass for the private context

profiles: *const AVProfile

< array of recognized profiles, or NULL if unknown, array is terminated by {FF_PROFILE_UNKNOWN}

priv_data_size: c_int next: *mut AVCodec init_thread_copy: Option<unsafe extern fn(_: *mut AVCodecContext) -> c_int>

@name Frame-level threading support functions @{ If defined, called on thread contexts when they are created. If the codec allocates writable tables in init(), re-allocate them here. priv_data will be set to a copy of the original.

update_thread_context: Option<unsafe extern fn(_: *mut AVCodecContext, _: *const AVCodecContext) -> c_int>

Copy necessary context variables from a previous thread context to the current one. If not defined, the next thread will start automatically; otherwise, the codec must call ff_thread_finish_setup().

dst and src will (rarely) point to the same context, in which case memcpy should be skipped.

defaults: *const AVCodecDefault

Private codec-specific defaults.

init_static_data: Option<unsafe extern fn(_: *mut AVCodec)>

Initialize codec static data, called from avcodec_register().

init: Option<unsafe extern fn(_: *mut AVCodecContext) -> c_int> encode_sub: Option<unsafe extern fn(_: *mut AVCodecContext, _: *mut u8, _: c_int, _: *const AVSubtitle) -> c_int> encode2: Option<unsafe extern fn(_: *mut AVCodecContext, _: *mut AVPacket, _: *const AVFrame, _: *mut c_int) -> c_int>

Encode data to an AVPacket.

@param avctx codec context @param avpkt output AVPacket (may contain a user-provided buffer) @param[in] frame AVFrame containing the raw data to be encoded @param[out] got_packet_ptr encoder sets to 0 or 1 to indicate that a non-empty packet was returned in avpkt. @return 0 on success, negative error code on failure

decode: Option<unsafe extern fn(_: *mut AVCodecContext, _: *mut c_void, _: *mut c_int, _: *mut AVPacket) -> c_int> close: Option<unsafe extern fn(_: *mut AVCodecContext) -> c_int> send_frame: Option<unsafe extern fn(_: *mut AVCodecContext, _: *const AVFrame) -> c_int>

Decode/encode API with decoupled packet/frame dataflow. The API is the same as the avcodec_ prefixed APIs (avcodec_send_frame() etc.), except that: - never called if the codec is closed or the wrong type, - AVPacket parameter change side data is applied right before calling AVCodec->send_packet, - if AV_CODEC_CAP_DELAY is not set, drain packets or frames are never sent, - only one drain packet is ever passed down (until the next flush()), - a drain AVPacket is always NULL (no need to check for avpkt->size).

send_packet: Option<unsafe extern fn(_: *mut AVCodecContext, _: *const AVPacket) -> c_int> receive_frame: Option<unsafe extern fn(_: *mut AVCodecContext, _: *mut AVFrame) -> c_int> receive_packet: Option<unsafe extern fn(_: *mut AVCodecContext, _: *mut AVPacket) -> c_int> flush: Option<unsafe extern fn(_: *mut AVCodecContext)>

Flush buffers. Will be called when seeking

caps_internal: c_int

Internal codec capabilities. See FF_CODEC_CAP_* in internal.h