Struct av_sys::AVHWAccel

#[repr(C)]

pub struct AVHWAccel {
    pub name: *const c_char,
    pub type_: AVMediaType,
    pub id: AVCodecID,
    pub pix_fmt: AVPixelFormat,
    pub capabilities: c_int,
    pub next: *mut AVHWAccel,
    pub alloc_frame: Option<unsafe extern fn(_: *mut AVCodecContext, _: *mut AVFrame) -> c_int>,
    pub start_frame: Option<unsafe extern fn(_: *mut AVCodecContext, _: *const u8, _: u32) -> c_int>,
    pub decode_slice: Option<unsafe extern fn(_: *mut AVCodecContext, _: *const u8, _: u32) -> c_int>,
    pub end_frame: Option<unsafe extern fn(_: *mut AVCodecContext) -> c_int>,
    pub frame_priv_data_size: c_int,
    pub decode_mb: Option<unsafe extern fn(_: *mut MpegEncContext)>,
    pub init: Option<unsafe extern fn(_: *mut AVCodecContext) -> c_int>,
    pub uninit: Option<unsafe extern fn(_: *mut AVCodecContext) -> c_int>,
    pub priv_data_size: c_int,
    pub caps_internal: c_int,
}

@defgroup lavc_hwaccel AVHWAccel @{

Fields
name: *const c_char

Name of the hardware accelerated codec. The name is globally unique among encoders and among decoders (but an encoder and a decoder can share the same name).

type_: AVMediaType

Type of codec implemented by the hardware accelerator.

See AVMEDIA_TYPE_xxx

id: AVCodecID

Codec implemented by the hardware accelerator.

See AV_CODEC_ID_xxx

pix_fmt: AVPixelFormat

Supported pixel format.

Only hardware accelerated formats are supported here.

capabilities: c_int

Hardware accelerated codec capabilities. see HWACCEL_CODEC_CAP_*

next: *mut AVHWAccel alloc_frame: Option<unsafe extern fn(_: *mut AVCodecContext, _: *mut AVFrame) -> c_int>

Allocate a custom buffer

start_frame: Option<unsafe extern fn(_: *mut AVCodecContext, _: *const u8, _: u32) -> c_int>

Called at the beginning of each frame or field picture.

Meaningful frame information (codec specific) is guaranteed to be parsed at this point. This function is mandatory.

Note that buf can be NULL along with buf_size set to 0. Otherwise, this means the whole frame is available at this point.

@param avctx the codec context @param buf the frame data buffer base @param buf_size the size of the frame in bytes @return zero if successful, a negative value otherwise

decode_slice: Option<unsafe extern fn(_: *mut AVCodecContext, _: *const u8, _: u32) -> c_int>

Callback for each slice.

Meaningful slice information (codec specific) is guaranteed to be parsed at this point. This function is mandatory. The only exception is XvMC, that works on MB level.

@param avctx the codec context @param buf the slice data buffer base @param buf_size the size of the slice in bytes @return zero if successful, a negative value otherwise

end_frame: Option<unsafe extern fn(_: *mut AVCodecContext) -> c_int>

Called at the end of each frame or field picture.

The whole picture is parsed at this point and can now be sent to the hardware accelerator. This function is mandatory.

@param avctx the codec context @return zero if successful, a negative value otherwise

frame_priv_data_size: c_int

Size of per-frame hardware accelerator private data.

Private data is allocated with av_mallocz() before AVCodecContext.get_buffer() and deallocated after AVCodecContext.release_buffer().

decode_mb: Option<unsafe extern fn(_: *mut MpegEncContext)>

Called for every Macroblock in a slice.

XvMC uses it to replace the ff_mpv_decode_mb(). Instead of decoding to raw picture, MB parameters are stored in an array provided by the video driver.

@param s the mpeg context

init: Option<unsafe extern fn(_: *mut AVCodecContext) -> c_int>

Initialize the hwaccel private data.

This will be called from ff_get_format(), after hwaccel and hwaccel_context are set and the hwaccel private data in AVCodecInternal is allocated.

uninit: Option<unsafe extern fn(_: *mut AVCodecContext) -> c_int>

Uninitialize the hwaccel private data.

This will be called from get_format() or avcodec_close(), after hwaccel and hwaccel_context are already uninitialized.

priv_data_size: c_int

Size of the private data to allocate in AVCodecInternal.hwaccel_priv_data.

caps_internal: c_int

Internal hwaccel capabilities.